Fix compile warnings from VC 2005
git-svn-id: https://svn.apache.org/repos/asf/xml/security/trunk@389736 13f79535-47bb-0310-9956-ffa450edef68

+++ c/src/dsig/DSIGAlgorithmHandlerDefault.cpp
@@ -458,7 +458,7 @@ unsigned int DSIGAlgorithmHandlerDefault::signToSafeBuffer(
 
 	result = b64Buf;
 
-	return strlen(b64Buf);
+	return (unsigned int) strlen(b64Buf);
 
 }
 
@@ -515,7 +515,7 @@ bool DSIGAlgorithmHandlerDefault::verifyBase64Signature(
 			hash, 
 			hashLen,
 			(char *) sig, 
-			strlen(sig));
+			(unsigned int) strlen(sig));
 
 		break;
 
@@ -533,7 +533,7 @@ bool DSIGAlgorithmHandlerDefault::verifyBase64Signature(
 			hash,
 			hashLen,
 			sig,
-			strlen(sig));
+			(unsigned int) strlen(sig));
 
 		break;
